requirements.txt
# Core
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
streamlit>=1.35.0
pydantic>=2.7.0

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9

# LLM
groq>=0.9.0

# Data
pandas>=2.2.0
numpy>=1.26.0
openpyxl>=3.1.2
plotly>=5.22.0

# Utils
python-dotenv>=1.0.1
requests>=2.31.0
